Proxy refactor
