+++ whproxy/errors.go
@@ -1,4 +1,4 @@
-package proxy
+package whproxy
 
 import (
 	"fmt"
