+++ whproxy/register-test
@@ -1,0 +1,1 @@
+session: proxy.go:86: worker with id validWorkerID registered on proxy
