+++ whproxy/request-test
@@ -1,0 +1,1 @@
+session: proxy.go:86: worker with id workerID registered on proxy
