+++ whproxy/wsproxy.go
@@ -1,4 +1,4 @@
-package proxy
+package whproxy
 
 import (
 	"net"
